crystalhd: add debug printk of register accesses
The CrystalHD device driver assumes a little endian host. In order to
verify future work on endianess-independence, this patch adds
printk's to track the register access to the device.

This helps verify bitfield endianess correctness as well as any
hard-coded assumptions in the driver code.

Signed-off-by: Leon Woestenberg <leon@sidebranch.com>

Minor formatting modifications and use of __func__ added by Jarod.

Signed-off-by: Jarod Wilson <jarod@redhat.com>