crystalhd: add locking around hw accesses
Fixed issue of holding HW lock and sleeping in kernel when probing for
picture  numbers. Fixed parsing of the Mode field so we can release DMA
buffers correctly. Correctly working with Adobe FP for both 70012    and
70015. Elimiated HW pause for 70012 since it is borked. Added SW pause
from apps side. Currently only for FP and GST. Need to
fix for other apps.

Signed-off-by: Narendra Sankar <nsankar@broadcom.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>