Revert "Assume b422mode = YUY2"
This reverts commit 66ec3a568f14981eff7ba1d876b2d30c14b39283 and replaces it
with a safer fix.

It looks like what happened here is that b422mode was originally a binary
flag, which was then repurposed to allow >2 values, but without its type
being changed to the correct enum:

    typedef enum _BC_OUTPUT_FORMAT {
            MODE420			= 0x0,
            MODE422_YUY2		= 0x1,
            MODE422_UYVY		= 0x2,
            OUTPUT_MODE420		= 0x0,
            OUTPUT_MODE422_YUY2	= 0x1,
            OUTPUT_MODE422_UYVY	= 0x2,
            OUTPUT_MODE420_NV12	= 0x0,
            OUTPUT_MODE_INVALID	= 0xFF,
    } BC_OUTPUT_FORMAT;

Rather than ignoring the UYVY possiblility, I changed the type of b422mode
to the correct enum.  There must be some combination of graphics hardware
for which this is required to make the colors in the video correct, although
I'd guess most CrystalHD users are using it with Intel GMA/HD graphics.